requirements.txt
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
matplotlib==3.9.2
numpy==2.0.2
pandas==2.2.3
pdfminer.six==20240706
plotly==5.24.1
seaborn==0.13.2
spacy==3.8.2
streamlit==1.41.1